"Closes #7.
Now if an attribute for a table passed to search does not conform to that tables schemas, the search does not proceed, and an error about this non conformance is shown. Great.

The error is still not specific to the manner of type validation failure, but it does say the error is about attribute errors. We could improve the error wording and even provide specific type failure errors in the manner of regular type validation, by further modifying jtype-system. But for now this is good.

Also a small improvement to runStoredAction improves the lack of clarity noted in #6, so these changes make progress on that issue too."